refactoring : Fuel에 전략패턴 적용
Fuel : 추상 클래스
RandomFuel : 구현 클래스
MockFuel : 테스트용 구현 클래스
