+++ src/main/java/step4/model/Car.java
@@ -1,4 +1,4 @@
-package step4;
+package step4.model;
 
 import java.util.Objects;
