+++ src/main/java/step4/model/Fuel.java
@@ -1,0 +1,9 @@
+package step4.model;
+
+import java.util.Random;
+
+public interface Fuel {
+
+    public boolean isValid();
+
+}
