+++ src/main/java/step4/model/Round.java
@@ -1,4 +1,6 @@
-package step4;
+package step4.model;
+
+import step4.util.Error;
 
 import java.util.Objects;
