+++ src/test/java/step4/CarTest.java
@@ -1,27 +1,32 @@
 package step4;
 
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import step4.mock.MockFuel;
+import step4.model.Car;
+import step4.model.Fuel;
 
 import static org.assertj.core.api.Assertions.*;
 
 class CarTest {
 
+    Car car;
+    Fuel fuel;
+
+    @BeforeEach
+    void setUp() {
+        car = Car.of("peter");
+        fuel = new MockFuel();
+    }
+
     @Test
     void create() {
-        Car car = Car.of("peter");
         assertThat(car).isEqualTo(Car.of("peter"));
     }
 
     @Test
     void move() {
-        Fuel fuel = new Fuel() {
-            @Override
-            public boolean isValid() {
-                return true;
-            }
-        };
-        Car peter = Car.of("peter");
-        peter.move(fuel);
-        assertThat(peter.getPosition()).isEqualTo(1);
+        car.move(fuel);
+        assertThat(car.getPosition()).isEqualTo(1);
     }
 }
