+++ src/test/java/step4/CarsTest.java
@@ -1,8 +1,12 @@
 package step4;
 
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
+import step4.mock.MockFuel;
+import step4.model.Car;
+import step4.model.Cars;
+import step4.model.Fuel;
+import step4.model.Result;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -18,12 +22,7 @@ public class CarsTest {
     void setUp() {
         cars = new Cars("peter,kassie,oak");
         result = getResult();
-        fuel = new Fuel() {
-            @Override
-            public boolean isValid() {
-                return true;
-            }
-        };
+        fuel = new MockFuel();
     }
 
     @Test
