+++ src/test/java/step4/RaceTest.java
@@ -2,6 +2,10 @@ package step4;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import step4.mock.MockFuel;
+import step4.model.Cars;
+import step4.model.Fuel;
+import step4.model.Race;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -14,12 +18,7 @@ public class RaceTest {
     void setUp() {
         cars = new Cars("peter,kassie,oak");
         race = new Race(cars, 3);
-        fuel = new Fuel() {
-            @Override
-            public boolean isValid() {
-                return true;
-            }
-        };
+        fuel = new MockFuel();
     }
 
     @Test
