+++ src/test/java/step4/RoundTest.java
@@ -1,9 +1,7 @@
 package step4;
 
 import org.junit.jupiter.api.Test;
-
-import java.util.ArrayList;
-import java.util.List;
+import step4.model.Round;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
