+++ src/test/java/step4/mock/MockFuel.java
@@ -1,0 +1,11 @@
+package step4.mock;
+
+import step4.model.Fuel;
+
+public class MockFuel implements Fuel {
+
+    @Override
+    public boolean isValid() {
+        return true;
+    }
+}
